Upload python script files again
